+++ typescript/modules/apps/base/workspace/angular-material/application-app/src/app/app.module.ts
@@ -43,7 +43,6 @@ import {
   // Config
   DateConfig,
   MediaConfig,
-  AuthenticationConfig,
   // Services
   CreateService,
   DisplayService,
@@ -53,12 +52,10 @@ import {
   MetaService,
   WorkspaceService,
   ErrorService,
-  AuthenticationSessionStoreService,
   DateStaticService,
   AllorsBarcodeKeypressService,
   AllorsFocusBehaviorSubjectService,
   RefreshBehaviorService,
-  AuthenticationService,
   DateService,
   AllorsFocusService,
   RefreshService,
@@ -67,7 +64,7 @@ import {
   MediaLocalService,
   AllorsDialogService,
   // Interceptors
-  AuthenticationInterceptor,
+  UnauthorizedInterceptor,
   // Directives
   AllorsFocusDirective,
   AllorsBarcodeDirective,
@@ -267,18 +264,8 @@ import { AppActionService } from './services/action.service';
   providers: [
     WorkspaceService,
     {
-      provide: AuthenticationService,
-      useClass: AuthenticationSessionStoreService,
-    },
-    {
-      provide: AuthenticationConfig,
-      useValue: {
-        url: environment.baseUrl + environment.authUrl,
-      },
-    },
-    {
       provide: HTTP_INTERCEPTORS,
-      useClass: AuthenticationInterceptor,
+      useClass: UnauthorizedInterceptor,
       multi: true,
     },
     { provide: AllorsBarcodeService, useClass: AllorsBarcodeKeypressService },